requirements.txt
python-gitlab==3.1.0
requests==2.26.0
urllib3==1.26.6